Begin work to account for MPC lower bound of zero
I didn't realize until now that it's possible for the limiting lower bound of the MPC to be zero, but cFunc is not degenerate.  CDC knew this (and had written it down), but I never noticed.  Code will behave badly when this is the case.

This commit begins to fix the Markov model, where I realized the problem can come up.  It also needs a fix in ConsIndShock, and possibly elsewhere.